requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-SocketIO==5.6.0
Flask-Limiter==3.5.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.0.post1
gunicorn
psycopg2-binary
python-socketio==5.16.0
eventlet>=0.35.2
bidict
gevent>=24.11.1
simple-websocket==1.0.0
requests==2.32.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# Payment Processing
razorpay==1.4.1

# Monitoring
sentry-sdk[flask]==1.40.0

# Email
Flask-Mail==0.9.1

# Cloud Storage
boto3==1.34.34

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
faker==22.0.0
